add %u handler

+++ libft/includes/libft.h
@@ -6,7 +6,7 @@
 /*   By: adenord <alexandre.denord@gmail.com>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/07/12 15:56:09 by adenord           #+#    #+#             */
-/*   Updated: 2023/07/28 17:08:45 by adenord          ###   ########.fr       */
+/*   Updated: 2023/07/29 11:04:36 by adenord          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -108,6 +108,6 @@ int			ft_strcmp(const char *s1, const char *s2);
 char		*get_next_line(int fd);
 void		ft_putstr(const char *str);
 void		ft_putchar(char c);
-void		ft_putnbr(int nb);
+void		ft_putnbr(long nb);
 
 #endif
